requirements.txt
# Dependencies
flask
langchain
pillow
google-generativeai
python-dotenv
langchain
transformers